// src/app/types/Clothing.tsx
export type ClothingArticle = {
  id: number;
  name: string;
  available: boolean;
  image: string;
  price: number;
  currency: string;
  description?: string;
  //   category: string;
  //   size: string[];
  //   color: string[];
  //   brand: string;
  //   rating: number;
  //   reviews: number;
  //   discount?: number; // Optional property
  //   stock?: number; // Optional property
  //   createdAt?: Date; // Optional property
  //   updatedAt?: Date; // Optional property
};